+++ x64barebones/Kernel/Syscalls/syscalls.c
@@ -44,17 +44,17 @@ static void sys_get_ticks(unsigned long long * ticks);
 static void sys_beep(uint32_t frequency);
 static void * sys_malloc(uint64_t size);
 static void sys_free(void * ptrToFree);
-static uint32_t sys_create_process_foreground(char* name, Function function, char **args, uint32_t parentPid, const int fileDescriptors[CANT_FILE_DESCRIPTORS]);
-static uint64_t sys_kill_process(uint32_t pid);
+static int32_t sys_create_process_foreground(char* name, Function function, char **args, uint32_t parentPid, const int fileDescriptors[CANT_FILE_DESCRIPTORS]);
+static int64_t sys_kill_process(uint32_t pid);
 static ProcessCopyList * sys_get_processes_copy();
 static uint32_t sys_get_pid();
 static uint32_t sys_get_parent_pid();
-static uint64_t sys_set_priority(uint32_t pid, uint64_t priority);
-static uint64_t sys_block(uint32_t pid);
-static uint64_t sys_waitpid(uint32_t pid);
+static int64_t sys_set_priority(uint32_t pid, uint64_t priority);
+static int64_t sys_block(uint32_t pid);
+static int sys_waitpid(uint32_t pid);
 static void sys_free_process_copy(ProcessCopyList * processCopyList);
-static uint32_t sys_create_process_background(char* name, Function function, char **args, uint32_t parentPid, const int fileDescriptors[CANT_FILE_DESCRIPTORS]);
-static uint64_t sys_get_pipe_id();
+static int32_t sys_create_process_background(char* name, Function function, char **args, uint32_t parentPid, const int fileDescriptors[CANT_FILE_DESCRIPTORS]);
+static int sys_get_pipe_id();
 static int16_t sys_pipe_open(int id, char mode);
 static int16_t sys_pipe_close(int id);
 static int16_t sys_pipe_write(int id, char* msg, int len);
@@ -179,7 +179,7 @@ uint64_t syscallsDispatcher(uint64_t rdi, uint64_t rsi, uint64_t rdx, uint64_t r
 // Syscall Read - ID = 0
 static void sys_read(uint8_t * buf, uint32_t count, uint32_t * readBytes) {
     uint32_t pid = getCurrentPid();
-    int64_t fd = getCurrentReadFileDescriptor();
+    int fd = getCurrentReadFileDescriptor();
 
     if(fd == STDIN){
         readFromKeyboard(buf, count, readBytes);
@@ -191,7 +191,7 @@ static void sys_read(uint8_t * buf, uint32_t count, uint32_t * readBytes) {
 // Syscall Write - ID = 1
 static void sys_write(uint8_t * buf, uint32_t * count) {
     uint32_t pid = getCurrentPid();
-    uint64_t fd = getCurrentWriteFileDescriptor();
+    int fd = getCurrentWriteFileDescriptor();
 
     if(fd == STDOUT || fd == STDERR){
         drawStringOnCursor(buf, count);
@@ -311,11 +311,11 @@ static void sys_free(void * ptrToFree) {
     return freeMemory(ptrToFree);
 }
 
-static uint32_t sys_create_process_foreground(char* name, Function function, char **args, uint32_t parentPid, const int fileDescriptors[CANT_FILE_DESCRIPTORS]){
+static int32_t sys_create_process_foreground(char* name, Function function, char **args, uint32_t parentPid, const int fileDescriptors[CANT_FILE_DESCRIPTORS]){
     return createProcessFromSched((char*) name, (char) 1, 3, (Function) function,(char**) args, (uint32_t)parentPid, 0, fileDescriptors);
 }
 
-static uint64_t sys_kill_process(uint32_t pid){
+static int64_t sys_kill_process(uint32_t pid){
     return killProcess(pid);
 }
 
@@ -331,15 +331,15 @@ static uint32_t sys_get_parent_pid(){
     return getCurrentParentPid();
 }
 
-static uint64_t sys_set_priority(uint32_t pid, uint64_t priority){
+static int64_t sys_set_priority(uint32_t pid, uint64_t priority){
     return setPriority(pid, priority);
 }
 
-static uint64_t sys_block(uint32_t pid){
+static int64_t sys_block(uint32_t pid){
     return block(pid);
 }
 
-static uint64_t sys_waitpid(uint32_t pid){
+static int sys_waitpid(uint32_t pid){
     return waitProcessPid(pid);
 }
 
@@ -347,11 +347,11 @@ static void sys_free_process_copy(ProcessCopyList * processCopyList){
     freeProcessCopy(processCopyList);
 }
 
-static uint32_t sys_create_process_background(char* name, Function function, char **args, uint32_t parentPid, const int fileDescriptors[CANT_FILE_DESCRIPTORS]) {
+static int32_t sys_create_process_background(char* name, Function function, char **args, uint32_t parentPid, const int fileDescriptors[CANT_FILE_DESCRIPTORS]) {
     return createProcessFromSched((char*) name, (char) 0, 3, (Function) function,(char**) args, (uint32_t)parentPid, 0, fileDescriptors);
 }
 
-static uint64_t sys_get_pipe_id(){
+static int sys_get_pipe_id(){
     return getPipeId();
 }
